decoupling services from EDI class
